RecyclerView corriendo bien
- Correccion en header
- Modificacion de ListaFragment.java
- Modificacion de MainActivity
- Correccion  de fragment_carrito.xml
