+++ app/src/main/res/layout/fragment_carrito.xml
@@ -9,6 +9,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:text="Este es el carrito" />
 
 </FrameLayout>
